travel_planner/backend/server/requirements.txt
Flask
Flask-Cors
google-generativeai
python-dotenv
 requests